# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Faculty Library Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface MessageResponse {
  message: string;
}

export type RegisterBodyRole =
  (typeof RegisterBodyRole)[keyof typeof RegisterBodyRole];

export const RegisterBodyRole = {
  STUDENT: "STUDENT",
  FACULTY: "FACULTY",
  LIBRARIAN: "LIBRARIAN",
  ADMIN: "ADMIN",
} as const;

export interface RegisterBody {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  role: RegisterBodyRole;
  department: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface UpdateProfileBody {
  name?: string;
  department?: string;
  phone?: string;
}

export interface ForgotPasswordBody {
  email: string;
}

export interface ResetPasswordBody {
  token: string;
  /** @minLength 6 */
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  STUDENT: "STUDENT",
  FACULTY: "FACULTY",
  LIBRARIAN: "LIBRARIAN",
  ADMIN: "ADMIN",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  department: string;
  phone?: string | null;
  isActive: boolean;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type UpdateUserBodyRole =
  (typeof UpdateUserBodyRole)[keyof typeof UpdateUserBodyRole];

export const UpdateUserBodyRole = {
  STUDENT: "STUDENT",
  FACULTY: "FACULTY",
  LIBRARIAN: "LIBRARIAN",
  ADMIN: "ADMIN",
} as const;

export interface UpdateUserBody {
  role?: UpdateUserBodyRole;
  isActive?: boolean;
}

export interface PaginatedUsers {
  data: User[];
  page: number;
  pageSize: number;
  totalRecords: number;
  totalPages: number;
}

export type BookFormat = (typeof BookFormat)[keyof typeof BookFormat];

export const BookFormat = {
  PHYSICAL: "PHYSICAL",
  DIGITAL: "DIGITAL",
} as const;

export interface Book {
  id: number;
  title: string;
  author: string;
  isbn: string;
  publisher: string;
  publicationYear: number;
  edition?: string | null;
  category: string;
  tags: string[];
  format: BookFormat;
  totalCopies: number;
  availableCopies: number;
  shelfLocation?: string | null;
  description?: string | null;
  coverImage?: string | null;
  createdAt: string;
}

export type CreateBookBodyFormat =
  (typeof CreateBookBodyFormat)[keyof typeof CreateBookBodyFormat];

export const CreateBookBodyFormat = {
  PHYSICAL: "PHYSICAL",
  DIGITAL: "DIGITAL",
} as const;

export interface CreateBookBody {
  title: string;
  author: string;
  isbn: string;
  publisher: string;
  publicationYear: number;
  edition?: string;
  category: string;
  tags?: string[];
  format: CreateBookBodyFormat;
  /** @minimum 1 */
  totalCopies: number;
  shelfLocation?: string;
  description?: string;
  coverImage?: string;
}

export type UpdateBookBodyFormat =
  (typeof UpdateBookBodyFormat)[keyof typeof UpdateBookBodyFormat];

export const UpdateBookBodyFormat = {
  PHYSICAL: "PHYSICAL",
  DIGITAL: "DIGITAL",
} as const;

export interface UpdateBookBody {
  title?: string;
  author?: string;
  isbn?: string;
  publisher?: string;
  publicationYear?: number;
  edition?: string;
  category?: string;
  tags?: string[];
  format?: UpdateBookBodyFormat;
  totalCopies?: number;
  shelfLocation?: string;
  description?: string;
  coverImage?: string;
}

export interface PaginatedBooks {
  data: Book[];
  page: number;
  pageSize: number;
  totalRecords: number;
  totalPages: number;
}

export interface ImportBooksBody {
  books: CreateBookBody[];
}

export interface ImportBooksResponse {
  imported: number;
  message: string;
}

export interface ImportRowError {
  row: number;
  field: string;
  reason: string;
}

export interface ImportErrorResponse {
  errors: ImportRowError[];
  message: string;
}

export type LoanStatus = (typeof LoanStatus)[keyof typeof LoanStatus];

export const LoanStatus = {
  ACTIVE: "ACTIVE",
  RETURNED: "RETURNED",
  OVERDUE: "OVERDUE",
} as const;

export interface Loan {
  id: number;
  userId: number;
  bookId: number;
  book?: Book;
  user?: User;
  borrowedAt: string;
  dueDate: string;
  returnedAt?: string | null;
  renewalsCount: number;
  status: LoanStatus;
}

export interface BorrowBookBody {
  bookId: number;
}

export interface PaginatedLoans {
  data: Loan[];
  page: number;
  pageSize: number;
  totalRecords: number;
  totalPages: number;
}

export interface MyLoansResponse {
  active: Loan[];
  history: Loan[];
}

export interface DashboardSummary {
  totalBooks: number;
  totalUsers: number;
  activeLoans: number;
  overdueLoans: number;
  availableBooks: number;
  myActiveLoans?: number | null;
  myOverdueLoans?: number | null;
}

export interface OverdueSummary {
  count: number;
  loans: Loan[];
}

export interface PopularBook {
  book: Book;
  borrowCount: number;
}

export interface PopularBooksResponse {
  data: PopularBook[];
}

export type ActivityItemAction =
  (typeof ActivityItemAction)[keyof typeof ActivityItemAction];

export const ActivityItemAction = {
  BORROWED: "BORROWED",
  RETURNED: "RETURNED",
  RENEWED: "RENEWED",
  OVERDUE: "OVERDUE",
} as const;

export interface ActivityItem {
  loan: Loan;
  action: ActivityItemAction;
  timestamp: string;
}

export interface RecentActivityResponse {
  data: ActivityItem[];
}

export type ListBooksParams = {
  /**
   * Search query (title, author, ISBN)
   */
  q?: string;
  category?: string;
  format?: ListBooksFormat;
  /**
   * Filter only books with available copies
   */
  available?: boolean;
  yearFrom?: number;
  yearTo?: number;
  page?: number;
  /**
   * @maximum 100
   */
  pageSize?: number;
};

export type ListBooksFormat =
  (typeof ListBooksFormat)[keyof typeof ListBooksFormat];

export const ListBooksFormat = {
  PHYSICAL: "PHYSICAL",
  DIGITAL: "DIGITAL",
} as const;

export type ListBookCategories200 = {
  categories: string[];
};

export type ListAllLoansParams = {
  userId?: number;
  bookId?: number;
  status?: ListAllLoansStatus;
  page?: number;
  /**
   * @maximum 100
   */
  pageSize?: number;
};

export type ListAllLoansStatus =
  (typeof ListAllLoansStatus)[keyof typeof ListAllLoansStatus];

export const ListAllLoansStatus = {
  ACTIVE: "ACTIVE",
  RETURNED: "RETURNED",
  OVERDUE: "OVERDUE",
} as const;

export type ListUsersParams = {
  page?: number;
  /**
   * @maximum 100
   */
  pageSize?: number;
  q?: string;
};

export type GetPopularBooksParams = {
  limit?: number;
};

export type GetRecentActivityParams = {
  limit?: number;
};