# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Faculty Library Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateMe
      tags: [auth]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordBody"
      responses:
        "200":
          description: Reset token sent (printed to console in dev)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordBody"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Books ──────────────────────────────────────────────────────────────────
  /books:
    get:
      operationId: listBooks
      tags: [books]
      summary: List and search books with filters and pagination
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Search query (title, author, ISBN)
        - name: category
          in: query
          schema:
            type: string
        - name: format
          in: query
          schema:
            type: string
            enum: [PHYSICAL, DIGITAL]
        - name: available
          in: query
          schema:
            type: boolean
          description: Filter only books with available copies
        - name: yearFrom
          in: query
          schema:
            type: integer
        - name: yearTo
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: Paginated books
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedBooks"

    post:
      operationId: createBook
      tags: [books]
      summary: Add a new book to the catalog (Librarian/Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookBody"
      responses:
        "201":
          description: Book created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /books/import:
    post:
      operationId: importBooks
      tags: [books]
      summary: Bulk import books via JSON array (Librarian/Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportBooksBody"
      responses:
        "201":
          description: Books imported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportBooksResponse"
        "422":
          description: Validation errors in import data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportErrorResponse"
...
  /books/{bookId}:
    get:
      operationId: getBook
      tags: [books]
      summary: Get a single book by ID
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Book
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateBook
      tags: [books]
      summary: Update a book (Librarian/Admin, cannot change availableCopies directly)
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookBody"
      responses:
        "200":
          description: Updated book
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"

    delete:
      operationId: deleteBook
      tags: [books]
      summary: Delete a book (only if no active loans)
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "409":
          description: Book has active loans
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Loans ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ─── Auth Schemas ───────────────────────────────────────────────────────
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [STUDENT, FACULTY, LIBRARIAN, ADMIN]
        department:
          type: string
      required: [name, email, password, role, department]
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    UpdateProfileBody:
      type: object
      properties:
        name:
          type: string
        department:
          type: string
        phone:
          type: string
    ForgotPasswordBody:
      type: object
      properties:
        email:
          type: string
          format: email
      required: [email]
    ResetPasswordBody:
      type: object
      properties:
        token:
          type: string
        password:
          type: string
          minLength: 6
      required: [token, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [STUDENT, FACULTY, LIBRARIAN, ADMIN]
        department:
          type: string
        phone:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, department, isActive, createdAt]
...
    Book:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        author:
          type: string
        isbn:
          type: string
        publisher:
          type: string
        publicationYear:
          type: integer
        edition:
          type: string
          nullable: true
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        format:
          type: string
          enum: [PHYSICAL, DIGITAL]
        totalCopies:
          type: integer
        availableCopies:
          type: integer
        shelfLocation:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        coverImage:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, title, author, isbn, publisher, publicationYear, category, tags, format, totalCopies, availableCopies, createdAt]
    CreateBookBody:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
        isbn:
          type: string
        publisher:
          type: string
        publicationYear:
          type: integer
        edition:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        format:
          type: string
          enum: [PHYSICAL, DIGITAL]
        totalCopies:
          type: integer
          minimum: 1
        shelfLocation:
          type: string
        description:
          type: string
        coverImage:
          type: string
      required: [title, author, isbn, publisher, publicationYear, category, format, totalCopies]
    UpdateBookBody:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
        isbn:
          type: string
        publisher:
          type: string
        publicationYear:
          type: integer
        edition:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        format:
          type: string
          enum: [PHYSICAL, DIGITAL]
        totalCopies:
          type: integer
        shelfLocation:
          type: string
        description:
          type: string
        coverImage:
          type: string
    PaginatedBooks:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Book"
        page:
          type: integer
        pageSize:
          type: integer
        totalRecords:
          type: integer
        totalPages:
          type: integer
      required: [data, page, pageSize, totalRecords, totalPages]
    ImportBooksBody:
      type: object
      properties:
        books:
          type: array
          items:
            $ref: "#/components/schemas/CreateBookBody"
      required: [books]
    ImportBooksResponse:
      type: object
      properties:
        imported:
          type: integer
        message:
          type: string
      required: [imported, message]
    ImportErrorResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ImportRowError"
        message:
          type: string
      required: [errors, message]
    ImportRowError:
      type: object
      properties:
        row:
          type: integer
        field:
          type: string
        reason:
          type: string
      required: [row, field, reason]

    # ─── Loan Schemas ───────────────────────────────────────────────────────
...